dbt-on-astro/requirements.txt
astronomer-cosmos==1.10.0
